evaluation/requirements.txt
ipython
pandas
requests
pyyaml
